# swagger.yaml
swagger: '2.0'
info:
  version: 1.0.0
  title: ILA-Server
host: ila.dev.kirschbaum.bayern
schemes:
  - https
securityDefinitions:
  jwt:
    type: apiKey
    name: Bearer
    in: header

tags:
  - name: Auth
  - name: Courses
  - name: Lectures

definitions:
  GenericErrorResponse:
    type: object
    properties:
      errors:
        type: array
        items:
          type: string

  AuthLoginResponse:
    type: object
    required:
      - token
    properties:
      token:
        type: string
  AuthLoginModel:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
      password:
        type: string
  AuthRegisterResponse:
    type: object
    properties:
      messages:
        type: array
        items:
          type: string
  AuthRegisterModel:
    type: object
    required:
      - email
      - password
      - name
    properties:
      email:
        type: string
      password:
        type: string
      name:
        type: string

  CoursesResponseModel:
    type: object
    properties:
      ownCourses:
        type: array
        items:
          $ref: '#/definitions/CourseModel'
      memberCourses:
        type: array
        items:
          $ref: '#/definitions/CourseModel'

  CourseModel:
    type: object
    required:
      - id
      - title
      - ownerId
    properties:
      id:
        type: integer
      title:
        type: string
      description:
        type: string
      ownerId:
        type: integer
      course_tokens:
        type: array
        items:
          $ref: '#/definitions/TokenModel'
      Members:
        type: array
        items:
          $ref: '#/definitions/MemberModel'
  MemberModel:
    type: object
    required:
      - id
      - name
      - email
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
  TokenModel:
    type: object
    required:
      - id
      - token
      - enabled
      - courseId
    properties:
      id:
        type: integer
      token:
        type: string
      enabled:
        type: boolean
      courseId:
        type: integer
  CourseCreateModel:
    type: object
    required:
      - title
    properties:
      title:
        type: string
      description:
        type: string
  CourseUpdateModel:
    type: object
    required:
      - id
      - title
    properties:
      id:
        type: integer
      title:
        type: string
      description:
        type: string

  LectureModel:
    type: object
    required:
      - id
      - topic
      - visible
      - courseId
    properties:
      id:
        type: integer
      topic:
        type: string
      description:
        type: string
      visible:
        type: boolean
      courseId:
        type: integer
  LectureCreateModel:
    type: object
    required:
      - topic
      - visible
      - courseId
    properties:
      topic:
        type: string
      description:
        type: string
      visible:
        type: boolean
      courseId:
        type: integer
  LectureUpdateModel:
    type: object
    required:
      - id
      - title
    properties:
      id:
        type: integer
      title:
        type: string
      description:
        type: string
  LectureVisibleModel:
    type: object
    required:
      - id
      - visible
    properties:
      id:
        type: integer
      visible:
        type: boolean


responses:
  InternalServerError:
    description: Internal Server Error
    schema:
      $ref: '#/definitions/GenericErrorResponse'
  Unauthorized:
    description: User is Unauthorized
  GenericError:
    description: A Generic Error Response
    schema:
      $ref: '#/definitions/GenericErrorResponse'

paths:
  /auth/login:
    post:
      summary: User login
      tags:
        - Auth
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/AuthLoginModel'
      responses:
        200:
          description: Returns the jwt
          schema:
            $ref: '#/definitions/AuthLoginResponse'
        400:
          $ref: '#/responses/GenericError'
        500:
          $ref: '#/responses/InternalServerError'
  /auth/register:
    post:
      summary: Users can register themselfs
      tags:
        - Auth
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/AuthRegisterModel'
      responses:
        200:
          description: Returns the jwt
          schema:
            $ref: '#/definitions/AuthRegisterResponse'
        400:
          $ref: '#/responses/GenericError'
        500:
          $ref: '#/responses/InternalServerError'

  /course:
    get:
      summary: Get a list of all your courses.
      tags:
        - Courses
      security:
        - jwt: []
      responses:
        200:
          description: Array of Courses
          schema:
            $ref: '#/definitions/CoursesResponseModel'
        500:
          $ref: '#/responses/InternalServerError'
    post:
      summary: Creates a new course.
      tags:
        - Courses
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/CourseCreateModel'
      responses:
        200:
          description: Returns the newly created course.
          schema:
            $ref: '#/definitions/CourseModel'
        400:
          $ref: '#/responses/GenericError'
        500:
          $ref: '#/responses/InternalServerError'
    put:
      summary: Update a course
      tags:
        - Courses
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/CourseUpdateModel'
      responses:
        200:
          description: Returns the updated course.
          schema:
            $ref: '#/definitions/CourseModel'
        400:
          $ref: '#/responses/GenericError'
        403:
          description: It's forbidden to update a foreign course.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        404:
          description: No course with the id found.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        500:
          $ref: '#/responses/InternalServerError'
    delete:
      summary: Delete one of your courses.
      tags:
        - Courses
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "id"
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Returns nothing.
          schema:
            type: object
        400:
          $ref: '#/responses/GenericError'
        403:
          description: It's forbidden to delete a foreign course.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        404:
          description: No course with the id found.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        500:
          $ref: '#/responses/InternalServerError'
  /course/generateToken:
    post:
      summary: Creates a new course.
      tags:
        - Courses
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: object
            properties:
              id:
                type: integer
                description: Id of the course.
      responses:
        200:
          description: Returns the newly created token.
          schema:
            $ref: '#/definitions/TokenModel'
        400:
          $ref: '#/responses/GenericError'
        403:
          description: It's forbidden to create a token for a foreign course.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        404:
          description: No course with the id found.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        500:
          $ref: '#/responses/InternalServerError'
  /course/join:
    post:
      description: Joins to a course
      tags:
        - Courses
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            type: object
            required:
              - courseId
              - token
            properties:
              courseId:
                type: integer
              token:
                type: string
      responses:
        200:
          description: Course joined
        400:
          $ref: '#/responses/GenericError'
        404:
          description: Couldn't find a course with the token.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        500:
          $ref: '#/responses/InternalServerError'
  /course/leave:
    delete:
      description: Leaves a course
      tags:
        - Courses
      security:
        - jwt: []
      responses:
        200:
          description: Course left.
        400:
          $ref: '#/responses/GenericError'
        404:
          description: Couldn't find a course.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        500:
          $ref: '#/responses/InternalServerError'

  /lecture:
    get:
      summary: Get a list of all lectures in the specific course.
      tags:
        - Lectures
      security:
        - jwt: []
      parameters:
        - in: path
          name: courseId
          type: integer
          required: true
      responses:
        200:
          description: Array of Lectures
          schema:
            type: array
            items:
              $ref: '#/definitions/LectureModel'
        400:
          $ref: '#/responses/GenericError'
        500:
          $ref: '#/responses/InternalServerError'
    post:
      summary: Creates a new lecture.
      tags:
        - Lectures
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/LectureCreateModel'
      responses:
        200:
          description: Returns the newly created lecture.
          schema:
            $ref: '#/definitions/LectureModel'
        400:
          $ref: '#/responses/GenericError'
        403:
          $ref: '#/responses/GenericError'
        404:
          $ref: '#/responses/GenericError'
        500:
          $ref: '#/responses/InternalServerError'
    put:
      summary: Update a lecture
      tags:
        - Lectures
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/LectureUpdateModel'
      responses:
        200:
          description: Returns the updated lecture.
          schema:
            $ref: '#/definitions/LectureModel'
        400:
          $ref: '#/responses/GenericError'
        403:
          description: It's forbidden to update a foreign courses lecture.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        404:
          description: No course with the id found.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        500:
          $ref: '#/responses/InternalServerError'
    delete:
      summary: Delete one of your lectures.
      tags:
        - Lectures
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "id"
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Returns nothing.
          schema:
            type: object
        400:
          $ref: '#/responses/GenericError'
        403:
          description: It's forbidden to delete a foreign courses lecture.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        404:
          description: No lecture with the id found.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        500:
          $ref: '#/responses/InternalServerError'
  /lecture/own:
    get:
      summary: Get a list of all lectures in a course where the user is the owner.
      tags:
        - Lectures
      security:
        - jwt: []
      parameters:
        - in: path
          name: courseId
          type: integer
          required: true
      responses:
        200:
          description: Array of Lectures
          schema:
            type: array
            items:
              $ref: '#/definitions/LectureModel'
        400:
          $ref: '#/responses/GenericError'
        500:
          $ref: '#/responses/InternalServerError'
  /lecture/visible:
    put:
      summary: Updates a lectures visibility
      tags:
        - Lectures
      security:
        - jwt: []
      parameters:
        - in: "body"
          name: "body"
          required: true
          schema:
            $ref: '#/definitions/LectureVisibleModel'
      responses:
        200:
          description: Returns the updated lecture.
          schema:
            $ref: '#/definitions/LectureModel'
        400:
          $ref: '#/responses/GenericError'
        403:
          description: It's forbidden to update a foreign courses lecture.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        404:
          description: No course with the id found.
          schema:
            $ref: '#/definitions/GenericErrorResponse'
        500:
          $ref: '#/responses/InternalServerError'